refactor(create_event_dialog): change file name
